fix:ランキングページ整え
