+++ src/components/Ranking.tsx
@@ -1,60 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { Link } from "react-router-dom";
-import { useTable, Column } from "react-table";
-
-type Data = {
-  rank: number;
-  name: string;
-  time: number;
-};
-
-const columns: Column<Data>[] = [
-  {
-    Header: "順位",
-    accessor: "rank",
-  },
-  {
-    Header: "名前",
-    accessor: "name",
-  },
-  {
-    Header: "記録",
-    accessor: "time",
-  },
-];
-
-const data: Data[] = [
-  {
-    rank: 1,
-    name: "John",
-    time: 23,
-  },
-  {
-    rank: 2,
-    name: "Jane",
-    time: 26,
-  },
-];
-
-const StyledTable = styled.div`
-  table {
-    border-collapse: collapse;
-    border-spacing: 0;
-    width: 100%;
-  }
-
-  tr {
-    border-bottom: solid 1px #eee;
-  }
-
-  th,
-  td {
-    text-align: center;
-    width: 25%;
-    padding: 10px 0;
-  }
-`;
+import Table from "./Table";
 
 const StyledDisplay = styled.div`
   min-width: 540px;
@@ -77,7 +24,7 @@ const StyledDiv = styled.div`
   background-repeat: no-repeat;
 `;
 
-const Header = styled.div`
+const StyledHeader = styled.div`
   display: grid;
   grid-template-columns: 50px 1fr 1fr;
   img {
@@ -99,57 +46,24 @@ const StyledLink = styled(Link)`
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
 `;
 
+const StyledPaging = styled.div`
+  margin: 0 auto;
+`;
+
 const Ranking: React.FC = () => {
-  const {
-    getTableProps,
-    getTableBodyProps,
-    headerGroups,
-    rows,
-    prepareRow,
-  } = useTable<Data>({ columns, data });
   return (
     <>
       <StyledDisplay>
         <StyledDiv>
-          <Header>
+          <StyledHeader>
             <StyledLink to="/">⬅</StyledLink>
             <img src="/images/trophy.png" alt="trophy" />
-          </Header>
-          <StyledTable>
-            <table {...getTableProps()}>
-              <thead>
-                {headerGroups.map((headerGroup) => (
-                  <tr {...headerGroup.getHeaderGroupProps()}>
-                    {headerGroup.headers.map((column) => (
-                      <th {...column.getHeaderProps()}>
-                        {column.render("Header")}
-                      </th>
-                    ))}
-                  </tr>
-                ))}
-              </thead>
-              <tbody {...getTableBodyProps()}>
-                {rows.map((row, i) => {
-                  prepareRow(row);
-                  return (
-                    <tr {...row.getRowProps()}>
-                      {row.cells.map((cell) => {
-                        return (
-                          <td {...cell.getCellProps()}>
-                            {cell.render("Cell")}
-                          </td>
-                        );
-                      })}
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
-          </StyledTable>
-          <div>
+          </StyledHeader>
+          <Table />
+          <StyledPaging>
             <button>←</button>
             <button>→</button>
-          </div>
+          </StyledPaging>
         </StyledDiv>
       </StyledDisplay>
     </>
